requirements.txt
dj-database-url==2.2.0
Django==4.2.3
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
python-decouple==3.8
whitenoise==6.7.0